reorganize audio scripts
