+++ VOCL_app/Assets/ConventionalAudio/MovingAverage.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 91ba3e3fb679f8b4e904ca50ce860a72
+guid: e7c1347fa31757b48a6fa1206c72d665
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
