+++ VOCL_app/Assets/Scenes/AudioTest.unity
@@ -271,7 +271,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 648074946}
-  - component: {fileID: 648074945}
   - component: {fileID: 648074947}
   - component: {fileID: 648074948}
   m_Layer: 0
@@ -281,25 +280,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &648074945
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 648074943}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1df4773fb37bf8040a8462482f2fba05, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_nRecordingHZ: 16000
-  m_level: 0
-  rmsLevel: 0
-  rmsLevelUnfilt: 0
-  m_5buff: 0
-  m_threshold: 0
-  detected: 0
 --- !u!4 &648074946
 Transform:
   m_ObjectHideFlags: 0
@@ -333,11 +313,6 @@ MonoBehaviour:
     m_PersistentCalls:
       m_Calls: []
   sampleRate: 0
-  m_level: 0
-  rmsLevel: 0
-  rmsLevelUnfilt: 0
-  m_5buff: 0
-  m_threshold: 0
 --- !u!114 &648074948
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -351,7 +326,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   done: 0
-  streamingMic: {fileID: 1553099040}
 --- !u!1 &1204480333
 GameObject:
   m_ObjectHideFlags: 0
@@ -435,55 +409,4 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1 &1553099039
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1553099041}
-  - component: {fileID: 1553099040}
-  m_Layer: 0
-  m_Name: streamingMic
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!114 &1553099040
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1553099039}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1df4773fb37bf8040a8462482f2fba05, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_nRecordingHZ: 16000
-  m_level: 0
-  rmsLevel: 0
-  rmsLevelUnfilt: 0
-  m_5buff: 0
-  m_threshold: 0
-  detected: 0
---- !u!4 &1553099041
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1553099039}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 659.5493, y: 181.63931, z: 722.6244}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
